+++ src/library/blas1/cldense-axpy.hpp
@@ -72,8 +72,8 @@ axpy(clsparse::array_base<T>& pR,
              << pY.data()
              << offset;
 
-    int blocksNum = (size + group_size - 1) / group_size;
-    int globalSize = blocksNum * group_size;
+    size_t blocksNum = (size + group_size - 1) / group_size;
+    size_t globalSize = blocksNum * group_size;
 
     cl::NDRange local(group_size);
     cl::NDRange global (globalSize);
